requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
clip-anytorch==2.5.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.2
fastapi==0.100.1
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gradio==3.35.2
gradio_client==0.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mobile-sam @ git+https://github.com/ChaoningZhang/MobileSAM.git@01ea8d0f5590082f0c1ceb0a3e2272593f20154b
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.1
opencv-python==4.8.0.74
orjson==3.9.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rpds-py==0.9.2
safetensors==0.3.1
scipy==1.11.1
seaborn==0.12.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
timm==0.9.2
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
ultralytics==8.0.120
urllib3==2.0.4
uvicorn==0.23.1
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
